fix: starting duplicate lazygits
**Issue:**

In some cases, when snacks is used to start a new lazygit, it will start
a new instance even if one is already open. This can result in multiple
lazygits being open for the same repository, wasting resources.

**Solution:**

Avoid doing this by caching the lazygits that are known to tsugit. In
addition:

- improve visibility of what is going on when developing and testing
  with a debug mode
- allow a repoLevelLazygit and a fileLevelLazygit to coexist. This
  resolves the following issue: when a fileLevelLazygit
  (`toggle_for_file()`) is started for the repository, and toggled off
  (hidden), it would remain open in the background. Now, when the
  fileLevelLazygit is toggled off, it will be hidden, and when it's
  closed, it will be closed permanently (exited and not restarted).

  The idea is to have a repolevelLazygit that is always open for the
  repository, and a fileLevelLazygit that is opened for a specific file
  for a short time inspection.