fix: displaying an empty lazygit after `force_quit`
**Issue:**

When lazygit is force-quit (e.g. by pressing `q`), toggling it back on
displays an empty lazygit pane.

Because of a race condition, the current way of reacting to the
`TermClose` event (i.e. when lazygit is closed) can sometimes be left
undone, which results in an invalid cache state.

**Solution:**

Wait until the buffer is created before setting up the `TermClose` event
handler. This ensures that the event handler is always set up, and we
can do cleanup properly.

Add tests to catch this issue in the future.