fix: only format long commit subjects specially when needed (#419)
I noticed over time that the special handling for commits that have long
subjects tends to be quite annoying to write because the auto hard
wrapping is disabled and I need to wrap my lines manually.

99% of the time I'm writing / editing commit messages that have a short
subject, so let's disable the special handling unless it's actually
needed to avoid these issues.