perf: only back up first 1000 lines of commit message
**Issue:**

With large refactorings and lockfiles, git commits can quite easily get
very large. I had a 4.2MB commit message as I use the `commit.verbose`
setting in git globally.

This causes a bit of lag when formatting the file on save, and lazygit
is somehow not able to save the changes to the file.

**Solution:**

Limit the number of lines backed up to 1000. This should be more than
enough.

It does not fully resolve the issue yet, but it's a good step in the
right direction.